Extract slice casting functions into a separate module
This is distinct functionality from the numeric traits that make up most of
`number.rs`.
