+++ src/lib.rs
@@ -105,6 +105,7 @@ mod model_metadata;
 mod number;
 mod op_registry;
 mod optimize;
+mod slice_cast;
 mod slice_reductions;
 mod tensor_pool;
 mod threading;
